Add semantic analysis phase
