+++ examples/09_type_error_string_plus_int.potion
@@ -1,0 +1,5 @@
+fn main() {
+    val age: int = 42
+    val message = "Age: " + age
+    print(message)
+}
