+++ examples/README.md
@@ -12,3 +12,4 @@ Files:
 - `06_end_to_end_features.potion` - a compact end-to-end example combining several language features.
 - `07_var_reassignment.potion` - sequential local reassignment with `var`.
 - `08_var_reassignment_control_flow.potion` - `var` reassignment across `if` and `match` branches.
+- `09_type_error_string_plus_int.potion` - intentionally invalid example showing compile-time rejection of `str + int`; use `to_string(...)` instead.
